changed UserController to an abstract class, checkChatRooms, checkMessages, checkContacts, and addContacts to abstract methods, and implemented AttendeeController
